[Optimize] move page_config to template_codec module
page_config should be moved to template_codec module, as it's more reasonable and maintainable.

issue: f-1128467656
AutoSubmit: true
