+++ core/renderer/build.gni
@@ -16,8 +16,6 @@ lynx_tasm_shared_sources = [
                              "lynx_env_config.h",
                              "lynx_global_pool.cc",
                              "lynx_global_pool.h",
-                             "page_config.cc",
-                             "page_config.h",
                              "page_proxy.cc",
                              "page_proxy.h",
                              "tasm/config.cc",
