+++ core/renderer/template_assembler.h
@@ -23,7 +23,6 @@
 #include "core/public/pipeline_option.h"
 #include "core/renderer/data/template_data.h"
 #include "core/renderer/events/touch_event_handler.h"
-#include "core/renderer/page_config.h"
 #include "core/renderer/page_proxy.h"
 #include "core/renderer/signal/signal_context.h"
 #include "core/renderer/tasm/i18n/i18n.h"
@@ -39,6 +38,7 @@
 #include "core/runtime/vm/lepus/vm_context.h"
 #include "core/services/timing_handler/timing.h"
 #include "core/template_bundle/lynx_template_bundle.h"
+#include "core/template_bundle/template_codec/binary_decoder/page_config.h"
 #include "core/template_bundle/template_codec/binary_decoder/template_binary_reader.h"
 #include "core/template_bundle/template_codec/moulds.h"
 #include "third_party/rapidjson/document.h"
