+++ core/renderer/template_entry.h
@@ -15,13 +15,13 @@
 #include "core/inspector/observer/inspector_lepus_observer.h"
 #include "core/renderer/css/css_style_sheet_manager.h"
 #include "core/renderer/data/template_data.h"
-#include "core/renderer/page_config.h"
 #include "core/renderer/tasm_runtime_bundle.h"
 #include "core/renderer/utils/base/element_template_info.h"
 #include "core/runtime/vm/lepus/lepus_global.h"
 #include "core/runtime/vm/lepus/vm_context.h"
 #include "core/template_bundle/lynx_template_bundle.h"
 #include "core/template_bundle/template_codec/binary_decoder/lynx_binary_lazy_reader_delegate.h"
+#include "core/template_bundle/template_codec/binary_decoder/page_config.h"
 #include "core/template_bundle/template_codec/binary_decoder/template_binary_reader.h"
 #include "core/template_bundle/template_codec/moulds.h"
