+++ core/shell/ios/tasm_platform_invoker_darwin.h
@@ -10,9 +10,9 @@
 #include <memory>
 #include <string>
 
-#include "core/renderer/page_config.h"
 #include "core/runtime/vm/lepus/lepus_value.h"
 #include "core/shell/tasm_platform_invoker.h"
+#include "core/template_bundle/template_codec/binary_decoder/page_config.h"
 #import "darwin/common/lynx/TemplateRenderCallbackProtocol.h"
 
 namespace lynx {
