+++ core/shell/tasm_platform_invoker.h
@@ -8,8 +8,8 @@
 #include <memory>
 #include <string>
 
-#include "core/renderer/page_config.h"
 #include "core/runtime/vm/lepus/lepus_value.h"
+#include "core/template_bundle/template_codec/binary_decoder/page_config.h"
 
 namespace lynx {
 namespace shell {
