+++ core/template_bundle/lynx_template_bundle.h
@@ -16,13 +16,13 @@
 
 #include "core/renderer/css/css_style_sheet_manager.h"
 #include "core/renderer/dom/element_bundle.h"
-#include "core/renderer/page_config.h"
 #include "core/renderer/template_themed.h"
 #include "core/renderer/utils/base/element_template_info.h"
 #include "core/runtime/piper/js/js_bundle.h"
 #include "core/runtime/vm/lepus/function.h"
 #include "core/runtime/vm/lepus/lepus_value.h"
 #include "core/runtime/vm/lepus/quick_context_pool.h"
+#include "core/template_bundle/template_codec/binary_decoder/page_config.h"
 #include "core/template_bundle/template_codec/binary_decoder/parallel_parse_task_scheduler.h"
 #include "core/template_bundle/template_codec/compile_options.h"
 #include "core/template_bundle/template_codec/header_ext_info.h"
