+++ core/template_bundle/template_codec/binary_decoder/BUILD.gn
@@ -21,6 +21,8 @@ binary_decoder_shared_sources = [
   "lynx_binary_lazy_reader_delegate.h",
   "lynx_binary_reader.cc",
   "lynx_binary_reader.h",
+  "page_config.cc",
+  "page_config.h",
   "parallel_parse_task_scheduler.cc",
   "parallel_parse_task_scheduler.h",
   "template_binary_reader.cc",
