+++ core/template_bundle/template_codec/binary_decoder/lynx_binary_config_helper.h
@@ -7,7 +7,7 @@
 
 #include <memory>
 
-#include "core/renderer/page_config.h"
+#include "core/template_bundle/template_codec/binary_decoder/page_config.h"
 #include "third_party/rapidjson/document.h"
 
 namespace lynx {
