+++ core/template_bundle/template_codec/binary_decoder/page_config.cc
@@ -2,7 +2,7 @@
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
 
-#include "core/renderer/page_config.h"
+#include "core/template_bundle/template_codec/binary_decoder/page_config.h"
 
 #include <vector>
