+++ platform/darwin/common/lynx/TemplateRenderCallbackProtocol.h
@@ -8,7 +8,7 @@
 #import <Foundation/Foundation.h>
 #import <Lynx/LynxTemplateBundle.h>
 #import "LynxErrorReceiverProtocol.h"
-#include "core/renderer/page_config.h"
+#include "core/template_bundle/template_codec/binary_decoder/page_config.h"
 
 namespace lynx {
 namespace tasm {
