+++ platform/darwin/ios/lynx/LUIConfigAdapter.h
@@ -3,7 +3,7 @@
 // LICENSE file in the root directory of this source tree.
 #import <Lynx/LUIConfig.h>
 
-#include "core/renderer/page_config.h"
+#include "core/template_bundle/template_codec/binary_decoder/page_config.h"
 
 @interface LUIConfigAdapter : NSObject <LUIConfig>
 - (instancetype)initWithConfig:(lynx::tasm::PageConfig*)pageConfig;
